created table component directory
