+++ src/page/Home/index.tsx
@@ -2,9 +2,9 @@ import { Center, Divider, Flex, VStack } from "@chakra-ui/react"
 
 import FrameContainer from "component/FrameContainer"
 import AccountPanel from "../../component/AccountPanel"
-import AccountInfoTable from "component/AccountInfoTable"
+import AccountInfoTable from "component/tables/AccountInfoTable"
 import TradeInfoPanel from "component/TradeInfoPanel"
-import OpenPositionsTable from "component/OpenPositionsTable"
+import OpenPositionsTable from "component/tables/OpenPositionsTable"
 import { PerpdexMarketContainer } from "../../container/connection/perpdexMarketContainer"
 import { PerpdexExchangeContainer } from "../../container/connection/perpdexExchangeContainer"
 import { User } from "../../container/connection/user"
